creating new models with address line 1,2, email, and phone
