+++ .openapi-generator/VERSION
@@ -1,1 +1,1 @@
-7.6.0
+7.7.0
