+++ README.md
@@ -1,16 +1,14 @@
 # Go API client for fideo
 
-This is a representation of the Fideo API based on the OpenAPI 3.0 specification.  You can find out more about
-Swagger at [https://swagger.io](https://swagger.io).
-Some useful links:
+Fideo Intelligence offers an identity intelligence product that protects the public good.
 - [Fideo Privacy Policy](https://www.fideo.ai/privacy-policy/)
 
 ## Overview
 This API client was generated by the [OpenAPI Generator](https://openapi-generator.tech) project.  By using the [OpenAPI-spec](https://www.openapis.org/) from a remote server, you can easily generate an API client.
 
-- API version: 1.0.0
-- Package version: 1.0.0
-- Generator version: 7.6.0
+- API version: 1.0.1
+- Package version: 1.0.1
+- Generator version: 7.7.0
 - Build package: org.openapitools.codegen.languages.GoClientCodegen
 
 ## Installation
