+++ go.mod
@@ -3,4 +3,12 @@ module github.com/GIT_USER_ID/GIT_REPO_ID
 go 1.18
 
 require (
+	github.com/stretchr/testify v1.9.0
+	gopkg.in/validator.v2 v2.0.1
+)
+
+require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
